Moved retrieve_none_or_403 to common.py
